+++ Source/SwiftLintFramework/Extensions/Array+SwiftLint.swift
@@ -59,7 +59,7 @@ extension Array {
         return parallelMap(transform: transform).compactMap { $0 }
     }
 
-    func parallelMap<T>(transform: (Element) -> T) -> [T] {
+    private func parallelMap<T>(transform: (Element) -> T) -> [T] {
         var result = ContiguousArray<T?>(repeating: nil, count: count)
         return result.withUnsafeMutableBufferPointer { buffer in
             DispatchQueue.concurrentPerform(iterations: buffer.count) { idx in